Convert Transmitter to a protocol
The goal here is to convert Transmitter to a proper protocol so one can create test
Transmitter instances. To do this a new class, G5Transmitter, is created which
implements the Transmitter protocol specific to a G5 transmitter (ie, bluetooth).

Making this change allows me to implement, for tests, a non-bluetooth dependent
Transmitter to greatly simplify my testing.