wait for 3 minutes before scanning again
In iOS versions < 11.2, the app waits for 2 seconds after reading from the transmitter before scanning again. Often the transmitter is still advertising, accepts the app's request to connect, but fails at the authentication step. Is there any reason why we can't use a 3 minute wait here, as we do for iOS 11.2 and above?