G6 Settings Page Date Formatting (sensor expiration timer) (#189)
* Settings Page Date Formatting

Settings Page Date Formatting (sensor expiration timer)

* Use setLocalizedDateFormatFromTemplate

Use setLocalizedDateFormatFromTemplate

* G6 sensor expiration display: Fix to date comparisons

A change in iOS17 changes the way relative date formatters return a date with a time. It no longer works in the comparison function. This change removes the time element from the comparison so that the logic works (works with the old way, too) and the correct time and date format are displayed.